third_party/blink/web_tests/virtual/coop-restrict-properties/external/wpt/html/cross-origin-opener-policy/tentative/restrict-properties/iframe-popup-to-so.https_1-2-expected.txt

This is a testharness.js-based test.
PASS COOP: restrict-properties to popup COOP: same-origin via an iframe, with same origin iframe, same origin popup with window_open
PASS COOP: restrict-properties to popup COOP: same-origin via an iframe, with same origin iframe, same origin popup with anchor
PASS COOP: restrict-properties to popup COOP: same-origin via an iframe, with same origin iframe, same origin popup with form
FAIL COOP: restrict-properties to popup COOP: same-origin via an iframe, with same site iframe, same origin popup with window_open assert_equals: Popup handle is non-null in iframe? expected "true" but got "false"
PASS COOP: restrict-properties to popup COOP: same-origin via an iframe, with same site iframe, same origin popup with anchor
PASS COOP: restrict-properties to popup COOP: same-origin via an iframe, with same site iframe, same origin popup with form
Harness: the test ran to completion.
